refactor: structured type guards and made them smaller and more readable
