+++ src/utils/is/primitives.ts
@@ -1,0 +1,1 @@
+export { isBoolean, isFunction, isNumber, isObject, isString, isUndefined } from 'radashi'
